img2img-turbo/requirements.txt
accelerate==0.34.2
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.2
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==3.30.3
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
diffusers==0.25.1
distlib==0.3.8
dominate==2.9.1
einops==0.8.0
exceptiongroup==1.2.1
fastapi==0.114.0
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
gradio==3.43.1
gradio_client==0.5.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-health-checking==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
identify==2.6.0
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.16
langchain-text-splitters==0.2.2
langsmith==0.1.85
lightning-utilities==0.11.7
lit==18.1.8
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy==1.11.1
mypy-extensions==1.0.0
narwhals==1.6.4
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open_clip_torch==2.26.1
openai==1.35.13
opencv-python==4.6.0.66
orjson==3.10.6
packaging==23.2
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
pinecone-client==4.1.2
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
pre-commit==3.8.0
protobuf==5.28.0
psutil==6.0.0
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.5
scipy==1.11.1
semantic-version==2.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.38.5
streamlit==1.38.0
streamlit-keyup==0.2.0
sympy==1.13.2
tenacity==8.5.0
timm==1.0.9
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchmetrics==1.4.1
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.35.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
validators==0.33.0
virtualenv==20.26.3
watchdog==4.0.2
wcwidth==0.2.13
weaviate-client==4.7.1
websockets==11.0.3
Werkzeug==3.0.3
xformers==0.0.22
yarl==1.9.4
zipp==3.20.1
